+++ app/controllers/TestController.ts
@@ -1,8 +1,8 @@
+import Redis from "@bejibun/redis";
 import {BunRequest} from "bun";
 import BaseController from "@/app/controllers/BaseController";
 import TestModel from "@/app/models/TestModel";
 import TestValidator from "@/app/validators/TestValidator";
-import Redis from "@/utils/Redis";
 
 export default class TestController extends BaseController {
     public async redis(request: BunRequest): Promise<Response> {