+++ utils/Enum.ts
@@ -1,7 +1,0 @@
-import EnumBuilder from "@/builders/EnumBuilder";
-
-export default class Enum {
-    public static setEnums(enums: any): EnumBuilder {
-        return new EnumBuilder(enums);
-    }
-}